Update patch set 1
Patch Set 1:

This patch fixes issue 1787 by removing the progress bar from the Music widgets. I was loath to do this, but frankly the android widget architecture is not designed to handle updates with such frequency (every second). As such, the previous progress bar code cheated by passing an incomplete RemoteViews object at each progress bar update interval during music playback. Thus, if the widget was rebuilt during playback (say, from a rotation of the device or if the launcher was restarted) then the widget lost all functionality EXCEPT for the progress bar.

One could re-instate the progress bar feature, but I think it would require a modification to the launcher to allow alternate means of updating widgets. (See, for example, some of the HTC sense implementations, I think.)

Also, I tried to make the minimum number of modifications so as to fix just this issue (and removing unnecessary code). There are several small visual issues with the Music app and widgets that I intend (hope?) to fix in a future update. I did fix one such issue with the 4x3 layout in this patch as well.

Patch-set: 1