Update patch set 2
Patch Set 2:

This is the response from the dev of ^3.

Sure I know you guys. This is awesome, from my side you can integrate and mod it at will. If you can contribute changes to the main tree that would be even cooler.

Cheers,
Filipe


So lets get the ball rolling!

Patch-set: 2