Update patch set 2
Patch Set 2:

I've been sitting on a bunch of small modifications to the music app since I figured you'd be making some drastic changes. I just wanted to warn you that you should be careful with most of the widgets from MusicMod - the code as it stood(stands?) was abusing the heck out of the binder by passing large bitmaps. I think if widgets with large artwork are desired, one needs to make some kind of contentprovider to serve the images. Would it be a terrible thing to take the base gingerbread player, and then add back in just the visual modifications (maybe with a few tweaks to the icons for style) and then separately the playback gestures? I'd be happy to help re-do the widgets, if that'd be helpful.

Patch-set: 2