Update patch set 1
Patch Set 1:

You tried setting dithering for the drawable, right?

http://developer.android.com/reference/android/graphics/drawable/Drawable.html#setDither%28boolean%29

This article talks about "choosing the correct format" which implies it can be defined explicitly from what I understand. They do it in the example code but I have not checked the rest of the code to see where that is done in this app.

http://www.curious-creature.org/2010/12/08/bitmap-quality-banding-and-dithering/

Patch-set: 1