Update patch set 1
Patch Set 1: No score

Yes, I was just about to ask what the resulting format would be after the compability changes added to make it work without the new drivers. I thought we could only render in 16bit without new drivers (though that the OpenGL driver on the other hand supported 32bit). Is that still true?

Still, I just examined the background during a call and that looked great as Jason mentioned so it should be fine here too as far as I can see. I did not find the equivilent of audio_player_background.xml in the files for the Phone app in cm github, that adds the gradient if I understood correctly when quickly looking through some of the files.

Patch-set: 1