Update patch set 1
Patch Set 1:

The code as it stands doesn't use the audio_player_background - I'm using that file as an alternate to generate the background via xml at home. (I just left it in the commit due to laziness.) The only background used is the bg_audio_player.9.png, FWIW. (This mirrors the phone app's setup for the call dialog).

I'm thinking of just going with the xml approach for now, as it's getting pretty close with my tweaking, and it has no banding issues.

Patch-set: 1