Update patch set 1
Patch Set 1:

Indeed, I stole only the HDPI version precisely for this reason. Also, on a side note, the phone app is missing the yellow hold version of the background in HDPI - they just have two boring grey ones. MDPI is ugly but has the colors correct. *sigh* :-)

I have the gradients done, and all that remains is to quickly photoshop some highlight backgrounds for the shuffle and repeat buttons (for trackball focus), and I'll upload the latest patch.

Patch-set: 1