Update patch set 1
Patch Set 1:

Yes, I noticed that. It seems the one in Froyo had it but back then it used the same version the mdpi drawables. 

https://github.com/CyanogenMod/android_packages_apps_Phone/blob/froyo/res/drawable-hdpi/bg_in_call_gradient_on_hold.9.png

I would submit a fixed version but I'm not sure what would happen to the 9-patch property of the image when it has passed Photoshop.

Sounds great. Looking forward to it.

Patch-set: 1