Update patch set 3
Patch Set 3:

Oh, I suppose you can get screenshots. Same link as last time:
http://www.dropbox.com/gallery/1969017/1/Music%20Player%20Screenshots?h=5652b8
The new xml gradient based backgrounds are pretty close to the png, now, after a bunch of tweaking. Also, I left the old shots in there for comparison - the new ones are those with the Arctic Monkey's track.
Also, if it's useful (and you can easily push this where you are), here is a link to the latest build from my home:
http://dl.dropbox.com/u/1969017/Music.apk

I found one tiny visual nit with the album info scrolling that I'll fix really quickly, but it won't change the screenshots, and I'll put a new apk in that dropbox folder/upload patch as soon as it's done.

Patch-set: 3