Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

Sorry for making you do that extra work. I just didn't turn on the box I build on so couldn't build it myself. Tried your apk though and it looks awesome. Only a few potential nitpicks/questions.

1. Is it supposed to be black at the bottom or gray? It's very obvious that it's not black when looking at it on an AMOLED. If it's intentional, nevermind. :)

2. Perhaps the selection color on the previous, play and next button should be green instead of orange. 

Other than the above, it's great. If the above stuff is intentional I would be fine with the way it is now. A definite +1.

Patch-set: 3