Update patch set 3
Patch Set 3:

Thanks for the feedback! I am looking for nits, so great! :-)

First, yes, the background is supposed to be grey at the bottom - it is a color match of the png, and it makes the screen look less flat.

On the second point, this was my first instinct too, and I have some screenshots with that option. Let me know if you like it better (last two):
http://www.dropbox.com/gallery/1969017/1/Music%20Player%20Screenshots?h=5652b8
I could even make them a little larger so they cover the active buttons a bit more.

Patch-set: 3