Update patch set 3
Patch Set 3:

Oh goodness - I gave you the opposite of what you asked for. (I just assumed you were thinking the same thing I was)

The standard android highlight is orange with a yellow click, so I made the bottom buttons match that. In the phone app they use green highlights/clicks for some buttons, so I opted to use that on the top. (The dialer app uses this sort of scheme, with green highlights for the numeric pad).

Do you want to see some green-styled button highlight shots?

Patch-set: 3