Update patch set 3
Patch Set 3:

Yes, I was think about what would look best. I just looked at the Dialer again though and you are right. If you want you can try it with green highlights and see if it looks better but I'll be fine with it if the current version gets submitted. :)

Patch-set: 3