Update patch set 3
Patch Set 3:

From the screenshot it looks like the play, formward, backwards buttons can be shrunk quite a bit for mdpi. Also, you can probably shrink the text below the album art or the album art itself unless it's the same size now as in the stock app. Changing the "Now playing" text kind of changes the whole appearence in my opinion. If that's the only good option though, so be it.

I agree with the rest of you about LDPI. Looking at how few drawables for LDPI exist in AOSP for newer versions shows how much Google cares about that size. I have to agree with them, it's just too small to be useful in most cases in my opinion. :(

Patch-set: 3