Update patch set 3
Patch Set 3:

Thanks for your input, Bjorn - I've been hemming and hawing over how to approach it. I've decided to give MDPI its own layout(s) so that whatever changes I make won't affect the HDPI setup.

I can indeed shrink the text by 2sp, and with some slight spacing adjustments all of the items might just fit on the page at MDPI, but it'll be tight. In landscape I can (and have) adjusted the album art size, but I feel that in portrait I liked the look of it lining up with the buttons below. However, I guess I have to decide which is the lesser evil: removing the 'Now playing' top, or shrinking the album art. I agree that making the top not a page title but an active song data display is undesirable in many ways. I don't really want to make the buttons smaller, as on many devices this could affect usability; I could shrink the play button in width, but not really in height.

Patch-set: 3