Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

It looks fine to me. With the different AR of MDPI the look can't be expected to be exactly the same. The difference in width between the album art and the buttons is big enough to not look like a mistake, but instead a well thought out decision. I'd say it's ready for merging. :)

Patch-set: 4