Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

(3 inline comments)

I'm sorry I've been incommunicado - I'm in the DC area and we have no power. I tested those MDPI layouts on an emulator yesterday afternoon, and the screenshot he showed looks like mine did in portrait. As long as landscape looks good too, I'm happy.

I have three whitespace nits in the comments above, and though I made a patch to fix these before we lost power, I was waiting for comments on MDPI before I uploaded them. But, now my fileserver is down, and I can't easily upload the change. If this is indeed in a state to be merged (and I hope it is), would someone be kind enough to remove those silly mistakes? If not, I'll do it as soon as we get power back, which should happen sometime in the next 24 hours. Thanks for all of your help and testing!

Patch-set: 4