Update patch set 4
Patch Set 4:

So we have power back, but it's a bit wonky. I was amending the silly whitespace fix changes when we lost it again, so now my local git information is totally corrupt. Do any of you know how to recover from this? *sigh* In the meantime, here is my latest build hot off of the presses:
http://dl.dropbox.com/u/1969017/Music.apk

Patch-set: 4