Update patch set 5
Patch Set 5: Looks good to me, but someone else must approve

The latest patch was just to fix the three whitespace issues, so now we should be golden.

James, I see now that when I build for nexus it only includes the hdpi layouts and so forth. Nothing other than whitespace has changed since this morning, so your build would still be an exceptionally accurate representation of the patch if one were to build it for mdpi. I'm sorry if I confused you with my link before - chalk it up to my inexperience.

Patch-set: 5