Update patch set 1
Patch Set 1:

I'm working on new 4x1 and 4x2 widgets, and I'll have screenshots in a little bit - the 4x1 is the same basic concept as the one from musicmod (albumart, info, controls) but I think it looks a little nicer. Also, it fixes some visual problems with the artwork/buttons and layout from the old mod.

For space's sake on the 4x1, how important is a previous button? I was also toying around with making the widgets have configurable components - just a thought, and it might not be worth the trouble.

Patch-set: 1