Update patch set 2
Patch Set 2:

Here are the promised new widgets, with screenshots:
http://www.dropbox.com/gallery/1969017/1/Music%20Player%20Widgets?h=96de5a

I'm using a slightly hack-y way of displaying the album artwork, as I'm passing along the appropriate URI (while checking for validity before passing it). Theoretically, this method could suffer from problems if, for example, someone removes the artwork during play (or something strange happens to the system-wide database). Of course the failure here would just mean that the widget would have an empty imageview.

If I were being more of an upstanding citizen, I'd write a little contentprovider to serve the album artwork, and if you guys think this is necessary, I'll do it. I should stress, though, that this current method is about a million times better than the old 'pass along a potentially huge bitmap through the binder' in that it will never kill the launcher nor cause binder errors.

Anyway, all comments are requested. :-)

Patch-set: 2