Update patch set 1
Patch Set 1:

If you like that better, I don't really mind - I thought the green was a bit nicer, but hey. :-) [I suppose if I really like it so much, I could just make it a theme option.]

On a side note - that progressbar style was included in the GB music player - I just decided to use it for the nowplaying screen. It was already in use in the streaming/preview window. Maybe the boys at Google considered using that as a style, just like they included some of the buttons from the gingerbread preview.

Is there something else that should be done to make the app play nicer with themes? As it stands, someone would have to change the new buttons and such.

Patch-set: 1