Update patch set 1
Patch Set 1:

Here is my take on gesture control for the music player. You can customize the gestures (and see a list of actions/defaults) by going to the Music settings and selecting 'Edit Gestures.' Long press on any entry for a customize context menu option. Custom gestures are only limited by the fact that in order to interact with the scrollable album information, they cannot simply be horizontal swipes (the editor should take care of this).

I'm especially looking for usability comments - if you hate any of the default gestures (and/or have better suggestions) please let me know.

Finally, now that we have a settings app it would be easy as pie to add an option for the new-and-now-defunct green GB-style progress bar.

Patch-set: 1