Update patch set 2
Patch Set 2:

In response to a request by Rick(fitsnugly), here is a feature from the last generation player.

During testing I noticed that there is a problem with notifications from Chrome to Phone (do any of you guys use this?) - they don't seem to trigger the standard transient audio focus loss (or any change in audio focus for that matter). Instead, you get a warning from the message queue that something (it looks like MediaPlayer.java@1386) is sending a message to a Handler on a dead thread. I haven't looked into the source, but I figured I'd mention it.

Also, if you want more features, just ask.

Patch-set: 2