Update patch set 2
Patch Set 2:

I could use some input here. The problem is that I am being greedy and ducking in response to AUDIOFOCUS_LOSS_TRANSIENT as well as the ..._CAN_DUCK events. This has the beneficial effect of having apps like navigation that don't declare their announcements as CAN_DUCK not interrupt playback for their very transient focus changes. Unfortunately, in a move that defies logic (at least to my naive senses), the phone app seems to only request transient focus for calls, although they may be of undetermined (long) duration. Thus, the music only ducks (and doesn't pause) with this patch in response to calls.

So, two choices: Make the ducking only for those which the system ordains as 'duckable' (and accept jerky playback during e.g. navigation) or change the phone app to request non-transient audio focus. Of course, the latter option will entail leaving the patch as-is, whereby it may have other unintended consequences for other 'transient' focus claims. Of course, the best move would be to change navigation to call its focus requests 'duckable,' but this is outside of our capacity.

Thoughts?

Patch-set: 2