Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

Thanks for the input - I did just that (implemented a phone state listener). The phone-related behavior is (during playback): if phone rings, pause playback - If call is dismissed, playback is automatically resumed. If call is answered playback stops, and the same thing happens when a call is placed. If the user is insane, and decides to put his/her caller on hold, open the Music player and presses play, the player will ignore the request.

Steve - don't you do some building on a mac? Do you know what's up with the libjpeg.so business?

Patch-set: 4