Update patch set 4
Patch Set 4: Verified; Looks good to me, approved

There's a patch in build that sorts the results of findleaves.py, which exposed a bunch of bugs in the build. I don't know why it's breaking prelinking of libjpeg on Mac, I haven't seen enough feedback about it. 

But no, I'm using Linux. Koush is drinking the Apple Kool-Aid.

Patch-set: 4