Exported MediaPlaybackService
As discussed in Issue 3215. I cant see any reason why this shouldnt be
set to exported=true. I dont think there is any security issues but
I coulndt find any description or reason for why Google set it to false.

With this its possible for Apps to fetch the currently playing song.

Change-Id: Ie559d61268f9d559cd28c727faa897a40fd6b1e7