Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

Since the media will be played through the A2DP sink, the hook for the BT connect message should watch for the A2DP sink to connect, not the BT headset profile.  This makes a difference for devices that offer an A2DP profile, but not AVRCP or headset.

Patch-set: 4