Update patch set 4
Patch Set 4:

Please remember that a A2DP device isn't always a headset. I use it as handsfree in my car and won't play audio with it. At least not automatically!

^^
Sorry, should be a zero score.
I just wantet to be sure you don't enable it by default.

Patch-set: 4